Create logic for playing rock paper scissors game
Create a function called getComputerChoice to choose a random move for
the "computer" player.

Create a function called playRound to play a single round of rock,
paper, scissors between a player and computer.

Create a function called game to play a best of five game of rock,
paper, scissors against a computer.